requirements.txt
langchain==0.16.3
langgraph==0.2.1
beautifulsoup4==4.11.1
requests==2.26.0
playwright==1.18.0
python-dotenv==0.19.2
tqdm==4.62.3
transformers==4.22.1
torch==1.10.2
pyqt5==5.15.6
